# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/service/pg-query": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Pg Query",
        "operationId": "pg_query_v1_service_pg_query_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PgQueryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PgQueryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/service/slack-post": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Slack Post",
        "operationId": "slack_post_v1_service_slack_post_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlackPostIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/service/agent-run": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Agent Run",
        "description": "Run one agent turn from an automation, reusing the exact reply paths the\nchat router uses (RAG-grounded, analytics text-to-SQL, or plain reasoning)\n\u2014 an automation must not get a second, differently-governed way to talk to\nan agent.",
        "operationId": "agent_run_v1_service_agent_run_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/service/email-send": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Email Send",
        "description": "Queue an email for human approval. **Never sends.**\n\n`email.send` is write:external (TRD \u00a79 names customer email as the\ncanonical always-approved case), so the only thing this endpoint can do is\ncreate the queue entry; the message goes on the wire in the approvals\nrouter's resume handler, after a person approves it.",
        "operationId": "email_send_v1_service_email_send_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailSendOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/service/notify": {
      "post": {
        "tags": [
          "service"
        ],
        "summary": "Notify",
        "description": "Internal notification: an audit-log entry an operator can find later.\n\nDeliberately the one recipe action with no outward effect at all \u2014 it is\nwhat a builder reaches for when the automation should leave a trace rather\nthan message a person.",
        "operationId": "notify_v1_service_notify_post",
        "parameters": [
          {
            "name": "x-fleet-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Fleet-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyIn"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Notify V1 Service Notify Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRunIn": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "question"
        ],
        "title": "AgentRunIn"
      },
      "AgentRunOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "text",
          "trace_id"
        ],
        "title": "AgentRunOut"
      },
...
      "EmailSendIn": {
        "properties": {
          "to": {
            "type": "string",
            "title": "To"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "to",
          "subject",
          "body"
        ],
        "title": "EmailSendIn"
      },
      "EmailSendOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "approval_id": {
            "type": "integer",
            "title": "Approval Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          }
        },
        "type": "object",
        "required": [
          "status",
          "approval_id",
          "run_id"
        ],
        "title": "EmailSendOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotifyIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "title",
          "message"
        ],
        "title": "NotifyIn"
      },
      "PgQueryIn": {
        "properties": {
          "sql": {
            "type": "string",
            "title": "Sql"
          }
        },
        "type": "object",
        "required": [
          "sql"
        ],
        "title": "PgQueryIn"
      },
      "PgQueryOut": {
        "properties": {
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          }
        },
        "type": "object",
        "required": [
          "rows",
          "row_count"
        ],
        "title": "PgQueryOut"
      },
...
      "SlackPostIn": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "text"
        ],
        "title": "SlackPostIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}